HV-732 Moving group conversion check into GroupConversionHelper
