+++ engine/src/main/java/org/hibernate/validator/internal/metadata/aggregated/GroupConversionHelper.java
@@ -20,10 +20,13 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+import javax.validation.GroupSequence;
 import javax.validation.metadata.GroupConversionDescriptor;
 
 import org.hibernate.validator.internal.metadata.descriptor.GroupConversionDescriptorImpl;
 import org.hibernate.validator.internal.metadata.facets.Cascadable;
+import org.hibernate.validator.internal.util.logging.Log;
+import org.hibernate.validator.internal.util.logging.LoggerFactory;
 
 import static org.hibernate.validator.internal.util.CollectionHelper.newHashSet;
 
@@ -33,6 +36,7 @@ import static org.hibernate.validator.internal.util.CollectionHelper.newHashSet;
  * @author Gunnar Morling
  */
 public class GroupConversionHelper {
+	private static final Log log = LoggerFactory.make();
 	private final Map<Class<?>, Class<?>> groupConversions;
 
 	public GroupConversionHelper(Map<Class<?>, Class<?>> groupConversions) {
@@ -74,5 +78,23 @@ public class GroupConversionHelper {
 		}
 
 		return Collections.unmodifiableSet( descriptors );
+	}
+
+	public void validateGroupConversions(Map<Class<?>, Class<?>> groupConversions, boolean isCascaded, String location) {
+		//group conversions may only be configured for cascadable elements
+		if ( !isCascaded && !groupConversions.isEmpty() ) {
+			throw log.getGroupConversionOnNonCascadingElementException( location );
+		}
+
+		//group conversions may not be configured using a sequence as source
+		for ( Class<?> oneGroup : groupConversions.keySet() ) {
+			if ( isGroupSequence( oneGroup ) ) {
+				throw log.getGroupConversionForSequenceException( oneGroup );
+			}
+		}
+	}
+
+	private boolean isGroupSequence(Class<?> oneGroup) {
+		return oneGroup.isAnnotationPresent( GroupSequence.class );
 	}
 }
